requirements.txt
crewai
crewai-tools
python-dotenv
litellm
phonikud
openai-whisper
edge-tts
pydub
playsound
websockets
fastapi
uvicorn
pytest
textblob
transformers
torch
scikit-learn
pandas
numpy
plotly
streamlit
matplotlib
seaborn
googletrans==4.0.0rc1
langdetect
spacy
speechbrain
librosa
protobuf
python-multipart
aiofiles
